Update to upstream version '1.8.2'
with Debian dir dd38b909d5eb94aa30944a98798f95de1147ad03

+++ src/selection.h
@@ -5,15 +5,15 @@
 
 typedef struct {
     BlockIter si;
-    long so;
-    long eo;
+    size_t so;
+    size_t eo;
     bool swapped;
 } SelectionInfo;
 
 void init_selection(const View *v, SelectionInfo *info);
-long prepare_selection(View *v);
-char *view_get_selection(View *v, long *size);
-int get_nr_selected_lines(const SelectionInfo *info);
-int get_nr_selected_chars(const SelectionInfo *info);
+size_t prepare_selection(View *v);
+char *view_get_selection(View *v, size_t *size);
+size_t get_nr_selected_lines(const SelectionInfo *info);
+size_t get_nr_selected_chars(const SelectionInfo *info);
 
 #endif
